Documents/capstone-Gurkirt/requirements.txt
pandas
numpy
fastapi
uvicorn
requests
python-dotenv
pytest